Moves spinup of server into lib.rs and creates first integration test.
