+++ Cargo.toml
@@ -4,6 +4,13 @@ version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path = "src/main.rs"
+name = "play_reading_party"
+
 
 [dependencies]
 actix-web = "4.3.1"
@@ -14,3 +21,8 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 
 dotenvy = "0.15.7"
+tokio = { version = "1", features = ["macros", "rt-multi-thread"]}
+
+
+[dev-dependencies]
+reqwest = "0.11"
